requirements.txt
# Install from repo root: pip install -r requirements.txt
# Core must be installed first so other modules can import it.
python-dotenv>=1.0.0
tushare>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
PyJWT>=2.7.0
pytest>=7.0.0
celery>=5.3.0
redis>=4.5.0
pydantic-settings>=2.0.0
httpx>=0.24.0
duckdb>=0.9.0
-e ./core
-e ./data-engine
-e ./data-pipeline
-e ./market-scanner
-e ./ai-models
-e ./feature-engine
-e ./backtest-engine
-e ./strategy
-e ./portfolio-engine
-e ./risk-engine
-e ./execution-engine
-e ./ai-lab
-e ./scheduler
-e ./evolution-engine
-e ./ai-fund-manager
-e ./alpha-factory
-e ./alpha-scoring
-e ./strategy-evolution
-e ./simulation-world
-e ./meta-fund-manager
-e ./gateway

# Optional / quant platform extras (init_data, parquet, unified logging)
yfinance>=0.2.40
pyarrow>=14.0.0
loguru>=0.7.0
pandas-datareader>=0.10.0
PyYAML>=6.0.1
vectorbt>=0.26.0
